requirements.txt
metadrive-simulator>=0.4.0
stable-baselines3[extra]>=2.0.0
tensorboard>=2.13.0
numpy>=1.24.0
